refactor: simplify types
